refactor(-): :hammer: move models motivos dieta
