+++ sme_sigpae_api/dados_comuns/__tests__/commands/test_carga_dados.py
@@ -20,6 +20,8 @@ from sme_sigpae_api.dieta_especial.models import (
     Alimento,
     AlimentoProprio,
     ClassificacaoDieta,
+)
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     MotivoAlteracaoUE,
     MotivoNegacao,
 )
