+++ sme_sigpae_api/dados_comuns/__tests__/test_metodos_carga_dados.py
@@ -16,6 +16,8 @@ from sme_sigpae_api.dieta_especial.models import (
     Alimento,
     AlimentoProprio,
     ClassificacaoDieta,
+)
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     MotivoAlteracaoUE,
     MotivoNegacao,
 )
