+++ sme_sigpae_api/dieta_especial/__tests__/conftest.py
@@ -15,10 +15,10 @@ from sme_sigpae_api.dieta_especial.models import (
     AlergiaIntolerancia,
     Alimento,
     ClassificacaoDieta,
-    MotivoNegacao,
 )
 from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     Anexo,
+    MotivoNegacao,
     SolicitacaoDietaEspecial,
 )
 from sme_sigpae_api.escola.models import Aluno, FaixaEtaria, PeriodoEscolar
