+++ sme_sigpae_api/dieta_especial/fixtures/factories/dieta_especial_base_factory.py
@@ -6,10 +6,12 @@ from sme_sigpae_api.dieta_especial.models import (
     AlergiaIntolerancia,
     Alimento,
     ClassificacaoDieta,
-    MotivoAlteracaoUE,
     ProtocoloPadraoDietaEspecial,
     SolicitacaoDietaEspecial,
     SubstituicaoAlimento,
+)
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    MotivoAlteracaoUE,
 )
 from sme_sigpae_api.escola.fixtures.factories.escola_factory import (
     AlunoFactory,
