+++ sme_sigpae_api/dieta_especial/models.py
@@ -27,32 +27,6 @@ class ClassificacaoDieta(Descritivel, Nomeavel):
         return self.nome
 
 
-class MotivoAlteracaoUE(Descritivel, Nomeavel, TemChaveExterna, Ativavel):
-    def __str__(self):
-        return self.nome
-
-    class Meta:
-        verbose_name = "Motivo Alteração U.E"
-        verbose_name_plural = "Motivo Alteração U.E"
-
-
-class MotivoNegacao(Descritivel):
-    CANCELAMENTO = "CANCELAMENTO"
-    INCLUSAO = "INCLUSAO"
-
-    PROCESSO_CHOICES = (
-        (CANCELAMENTO, "Solicitação de Cancelamento"),
-        (INCLUSAO, "Solicitação de Inclusão"),
-    )
-
-    processo = models.CharField(
-        choices=PROCESSO_CHOICES, default=INCLUSAO, blank=False, max_length=20
-    )
-
-    def __str__(self):
-        return self.descricao
-
-
 class SolicitacoesDietaEspecialAtivasInativasPorAluno(models.Model):
     aluno = models.OneToOneField(Aluno, on_delete=models.DO_NOTHING, primary_key=True)
     ativas = models.IntegerField()
